requirements.txt
streamlit
langchain
chromadb
openai
PyMuPDF
python-docx
markdown
PyYAML
requests
beautifulsoup4 